tests/requirements.txt
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.1
unittest-mock>=1.5.0
google-generativeai>=0.7.2
python-dotenv>=1.0.0